+++ Software/Photonfirmware/AnimatronicEyesTest/src/TPPAnimateServo.h
@@ -41,11 +41,11 @@ class TPP_AnimateServo{
         TPP_AnimateServo();
         void begin(int servoNum, int postion);
         void process(); // called every time in the loop to keep the eyes moving
-        int moveTo (int newX, int speed);
+        int moveTo (int newX, float speed);
 
     private:
+        
         static void initPWM();      // called once in the class inititator to init pwm library
-
         int servoNum_ = 0;          // Number of this servo on the driver board 
         float position_ = -1;       // the current position of the servo
         int destination_ = 0;       // the position we are heading towards
